feat(spinner): style spinner with currentColor and Box
